requirements.txt
pyautogen
openai
google-genai
groq
transformers
torch
python-dotenv
autogen-ext[openai]